+++ CHANGELOG.md
@@ -7,7 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-## [v0.2.0] - 2019-08-29
+## [v0.1.2] - 2019-11-22
+
+### Added
+
+- Support for v0.13.x of `generic-array`, in addition to the existing support
+  for v0.12.x of `generic-array`.
+
+## [v0.2.0] - 2019-08-29 - YANKED
 
 - The 0.1.1 release was a breaking change, now in 0.2.0
 
@@ -19,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Initial release
 
-[Unreleased]: https://github.com/japaric/as-slice/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/japaric/as-slice/compare/v0.1.2...HEAD
+[v0.1.2]: https://github.com/japaric/as-slice/compare/v0.1.1...v0.1.2
 [v0.2.0]: https://github.com/japaric/as-slice/compare/v0.1.1...v0.2.0
 [v0.1.1]: https://github.com/japaric/as-slice/compare/v0.1.0...v0.1.1